Fix: Update autocommit.sh permissions
This commit updates the permissions of the autocommit.sh script so that it is executable.

The old permissions (100644) did not allow the script to be executed, which prevented it from being used in the automated build process. The new permissions (100755) allow the script to be executed, which ensures that the automated build process can run successfully.